+++ src/component/Blog/index.ts
@@ -1,0 +1,1 @@
+export { BlogSection } from "./BlogSection";
