+++ src/component/Portfolio/index.ts
@@ -1,0 +1,1 @@
+export { PortfolioSection } from "./PortfolioSection";
